// src/app/components/lit-author-view/interfaces/lit-author-view.interface.ts
export interface ILitAuthorView {
      authorId: number;    // System.Int32
      firstName: string | null;    // System.String
      lastName: string | null;    // System.String
      birthDate: number | null;    // System.DateTime ?
      deathDate: number | null;    // System.DateTime ?
      iso3CntrRef: string | null;    // System.String
      iso2CntrRef: string | null;    // System.String
      cCountryName: string | null;    // System.String
}